Updated matrix lib
added matrix3 math/logical operators and methods. I'd advise anyone to not mess with this lib much yet because there are some things that aren't working right now, such as a couple of constructors, but the base is there and all the methods should work. Will fix the issues ASAP.